Refactor: Refactor the Components , Utils ,Constant and pages
